Create PersonAuthLink mechanism
This is a system for authenticating users by generating a link containing a
secret, emailing it to them, and considering them authenticated if they
received it. There's no UI for creating or using such links yet, but that'll
come.

Also, set the default timezone to America/New_York, which is presumably better
for our audience than UTC.