Associate PersonComment with User, not Person
We now expect to frequently have a User object (because Google and MIT
login is going to be common), and saving the User makes more sense than
the Person. (Ideally, we'd default in the admin to "the logged in user",
but I don't know how to, so we'll just trust people to fill it in.)

Also, add some grad_year help text.