Add PersonFrequency field
This is only sort of a property of the *member*, but it's maybe
generally useful, and we'll want it when we start using the DB for gate.
If we're using the DB to generate paper gate sheets, we'll really want
it, and if we do electronic gate, it's probably still useful for
deciding the order to show people. (I tentatively figure the first group
of names will be subscriptions + frequent attendees. Eventually we'll be
able to figure out frequent attendees from the actual attendance
records, but the value from the old DB will help bootstrap.)