signin: Distinguish in-flight and successful requests
This changes the messages in the signin UI from immediately saying "has
been marked", to starting out as "is being marked" and only changing to
"has been marked" when the request to the server completes successfully.
Apparently the success callback only triggers when the request actually
succeeds[1], so there's more work to be done to mark *failed* updates
usefully.

[1] https://api.jquery.com/jQuery.post/