signin: Add experimental table-based payment display
I don't love the current bootstrap alert based display of payments --
there's a fair amount of complexity in figuring out the appropriate
message and it seems not very skimmable.

This adds a prototype table based display -- there's columns for most of
the information that you might want, and I think it's easier to skim and
see what happened. I have not yet integrated it with the AJAX
success/failure actions, so it only displays live data in the final,
successful format. There are three sample rows to show what a failed
update, one still submitting, and a success would look like, though.