signin: Prototype split-dropdown for next-term subscriptions
Currently, people who get into the dance for free (MIT students or
people who bought subscriptions for the current term) can't buy a
subscription for the next dance (or do anything else beyond being marked
present). This commit adds two prototypes using Bootstrap's
split-dropdown support for having the usual dropdown as well. It
noticeably increases button width and requires non-trivial dexterity to
hit the button on the tablet, so I don't love it. Also the first
prototype, using the emoji to trigger the dropdown, is a hack and I
imagine later Bootstrap versions could break it, plus it breaks the
rounded corners, so while it's a little narrower than the second
prototype, it has other issues.

My inclination remains to add a checkbox that makes every button the
usual dropdown -- no additional width on weeks we're not selling lots of
subscriptions, but when we are you get a full-size button to click.