gate: Switch to split buttons for subscription payments
Preferences were fairly mixed between the styles, but it was Luke's top
and Ginda's second, and the other three were in at least one of their
bottom two. (I think I prefer the checkbox or longpress, but I guess
that's weird -- both had longpress last, and the checkbox was second or
third.) This might also be the least fiddly to get working / least
hackish, so that's another advantage. (It's even got
rounded corners without special effort.)