Finish offline-dev support
Developing without internet (eg, on a plane) is sometimes convenient,
but we use several Javascript CDNs, so it requires using different
locations for our JS libs. For now switching between offline and online
is a matter of moving a couple end comment markers around -- making it a
setting instead would be nice, but seems slightly tricky.