Pin to Django<4
We have a few incompatibilities with Django 4, so right now we *can't*
use Django 4.0. More importantly, Django 3.2 is an LTS supported through
April 2024, while Django 4.0 is only supported through April 2023.
Absent compelling reason, I think we'd rather be on an LTS so we're not
forced to upgrade for another year.