gate: Bulk add subscriptions
Add a UI for adding many subscriptions at once. The current expectation
is that for resolving the Spring 2020 stuff, I'll use the squares-pay
importer for the squares-pay folks, and then we look at who has their
name crossed off vs. who's marked in green (MIT+subscriber) and use this
for the rest.