gate: Display payment and sub status on voting members page
This makes the voting member page essentially the same as a filled-in
paper gate sheet, except only containing attendees and sorted by date.
Turning it into a backup gate sheet for in case the DB breaks is likely
not a lot of additional work.

The display is currently not ideal, due to being written without
internet (or local CSS docs):
- Ideally, the dance dates would be vertical, or otherwise formatted to
  take less horizontal space
- The current subscription status would ideally use a whole-cell
  strikethrough
- The green for the present "X" seems decidedly dubious