membership: Move "Update" button on profile page closer to top
This will hopefully make it easier to find, now that this page is
getting fairly long. The page has essentially three sections --
self-editable, admin-editable, and mailing lists. Putting the edit
button at the bottom of the self-editable section probably makes sense
and is easier to find than putting it between admin-editable and mailing
lists, where it's out of sight (depending on screen size) when making
changes, and also not found by quickly scrolling to the top and bottom.
(This isn't purely theoretical -- I watched somebody try to edit their
profile and apparently struggle to find the edit button.)