install: Tweaks to the Ansible installer and dev docs
Ansible doesn't appear to actually need changes at the moment, but will
in the future, and the error message was confusing. (The actual error
was because commit c329588b8cbf hadn't been merged to main yet.)

In the docs, people installing to do development probably want the `dev`
extra (mypy and pylint).