mypy: Don't use lazy ManyToManyField with implicit app label
In newer mypy[1][2], `ManyToManyField`s with lazy (string) models but no
app label don't work. For the `to` model, we can just use `Person`
directly (and do on the line above), so do that. For the `through`
model, add the app label -- Django doesn't care, and it makes mypy
happy.

Also, in mypy, well-structured wildcard patterns match without a
submodule too, so simplify some config.

[1] https://github.com/typeddjango/django-stubs/issues/1802#issuecomment-1781423719,
[2] https://github.com/typeddjango/django-stubs/pull/1719#issuecomment-1738691940
[3] https://mypy.readthedocs.io/en/stable/config_file.html#config-file-format