bootstrap: Switch to using local, pip-managed bootstrap
Bootstrap from a CDN requires internet, which is sometimes annoying for
development. (Also, AIUI it's falling out of fashion, due to privacy or
connection reuse or something.) Previously we used Bootstrap symlinked
in from Ubuntu packages, but that's problematic with non-Ubuntu devs.

Apparently Bootstrap has also been packaged in the pip ecosystem for
Django users, which seems like a good fit for us.