pay: Client-side validation of positive amount
It's somewhat tricky to validate *server-side* that the payment amount
is positive before the transaction hits the DB (we'd need to validate
two formsets against each other and I'm not sure that works), but we
already have the infrastructure to calculate the total *client-side*, so
it's easy to enforce on the client. That seems good enough -- I believe
Cybersource will validate it's a positive amount before charging the
user, and a zero-amount purchase isn't really a problem in the DB
anyway.