gate: Use `date` widgets on the new subscription page
This gives a handy browser widget to enter the date (so, e.g., it's easy
to say "it's the second Tuesday in September" and find that on a
calendar, rather than needing to know the specific day of the month).

I'm not sure why this isn't the default for Django.